feat(db): models crate
