+++ crates/storage/db-api/Cargo.toml
@@ -14,11 +14,12 @@ workspace = true
 [dependencies]
 # reth
 reth-codecs.workspace = true
+reth-db-models.workspace = true
 reth-primitives = { workspace = true, features = ["reth-codec"] }
 reth-primitives-traits.workspace = true
 reth-prune-types.workspace = true
-reth-storage-errors.workspace = true
 reth-stages-types.workspace = true
+reth-storage-errors.workspace = true
 reth-trie-common.workspace = true
 
 # codecs
