+++ crates/storage/db-api/src/models/accounts.rs
@@ -7,55 +7,8 @@ use crate::{
     table::{Decode, Encode},
     DatabaseError,
 };
-use reth_codecs::{derive_arbitrary, Compact};
-use reth_primitives::{Account, Address, BlockNumber, Buf, StorageKey};
+use reth_primitives::{Address, BlockNumber, StorageKey};
 use serde::{Deserialize, Serialize};
-
-/// Account as it is saved in the database.
-///
-/// [`Address`] is the subkey.
-#[derive_arbitrary(compact)]
-#[derive(Debug, Default, Clone, Eq, PartialEq, Serialize)]
-pub struct AccountBeforeTx {
-    /// Address for the account. Acts as `DupSort::SubKey`.
-    pub address: Address,
-    /// Account state before the transaction.
-    pub info: Option<Account>,
-}
-
-// NOTE: Removing reth_codec and manually encode subkey
-// and compress second part of the value. If we have compression
-// over whole value (Even SubKey) that would mess up fetching of values with seek_by_key_subkey
-impl Compact for AccountBeforeTx {
-    fn to_compact<B>(&self, buf: &mut B) -> usize
-    where
-        B: bytes::BufMut + AsMut<[u8]>,
-    {
-        // for now put full bytes and later compress it.
-        buf.put_slice(self.address.as_slice());
-
-        let mut acc_len = 0;
-        if let Some(account) = self.info {
-            acc_len = account.to_compact(buf);
-        }
-        acc_len + 20
-    }
-
-    fn from_compact(mut buf: &[u8], len: usize) -> (Self, &[u8]) {
-        let address = Address::from_slice(&buf[..20]);
-        buf.advance(20);
-
-        let info = if len - 20 > 0 {
-            let (acc, advanced_buf) = Account::from_compact(buf, len - 20);
-            buf = advanced_buf;
-            Some(acc)
-        } else {
-            None
-        };
-
-        (Self { address, info }, buf)
-    }
-}
 
 /// [`BlockNumber`] concatenated with [`Address`].
 ///
