+++ crates/storage/db-api/src/models/mod.rs
@@ -21,6 +21,7 @@ pub mod storage_sharded_key;
 pub use accounts::*;
 pub use blocks::*;
 pub use client_version::ClientVersion;
+pub use reth_db_models::{AccountBeforeTx, StoredBlockBodyIndices};
 pub use sharded_key::ShardedKey;
 
 /// Macro that implements [`Encode`] and [`Decode`] for uint types.
