+++ crates/storage/db-models/Cargo.toml
@@ -17,6 +17,7 @@ reth-codecs.workspace = true
 reth-primitives = { workspace = true, features = ["reth-codec"] }
 
 # codecs
+modular-bitfield.workspace = true
 serde = { workspace = true, default-features = false }
 
 # misc
@@ -27,13 +28,14 @@ arbitrary = { workspace = true, features = ["derive"], optional = true }
 proptest = { workspace = true, optional = true }
 
 [dev-dependencies]
-# reth libs with arbitrary
+# reth
 reth-primitives = { workspace = true, features = ["arbitrary"] }
 reth-codecs.workspace = true
 
 arbitrary = { workspace = true, features = ["derive"] }
-proptest.workspace = true
 proptest-arbitrary-interop.workspace = true
+proptest.workspace = true
+test-fuzz.workspace = true
 
 [features]
 test-utils = ["arbitrary"]
