+++ crates/storage/db-models/src/lib.rs
@@ -1,4 +1,9 @@
 //! Models used in storage module
 
-/// Database models
-pub mod models;
+/// Accounts
+pub mod accounts;
+pub use accounts::AccountBeforeTx;
+
+/// Blocks
+pub mod blocks;
+pub use blocks::StoredBlockBodyIndices;
