--- crates/storage/db-models/src/models/mod.rs
@@ -1,5 +1,0 @@
-/// Account
-pub use accounts::AccountBeforeTx;
-
-/// Accounts
-pub mod accounts;
