+++ crates/storage/storage-api/Cargo.toml
@@ -14,8 +14,8 @@ workspace = true
 [dependencies]
 # reth
 reth-chainspec.workspace = true
+reth-db-models.workspace = true
 reth-execution-types.workspace = true
-reth-db-api.workspace = true
 reth-primitives.workspace = true
 reth-prune-types.workspace = true
 reth-stages-types.workspace = true
